app/requirements.txt
streamlit
rpy2
pandas
plotly
seaborn
pillow